switch bfloat to float

+++ train_kaituoxu.py
@@ -343,7 +343,7 @@ def fit(model, train_set, test_set, criterion, optimizer, lr_updater, epochs, hi
 
 def forward(model, x, y, signal_length, criterion, device):
     if CFG["device"] == "cuda":
-        with torch.cuda.amp.autocast(dtype=torch.bfloat16):
+        with torch.cuda.amp.autocast(dtype=torch.float16):
             output = model(x)
 
             if CFG["loss"] == "si_snr":
